==> 2.DescriptiveAnalysis.Rmd <==
---
title: "Project_descriptive_analysis"
date: "2023-12-04"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# Here's the R workspace link for this file:
# https://drive.google.com/file/d/1Yvzm8HzQ9NLTax_ZyZN0CTfkatoYf9SS/view?usp=sharing
# model_table=df_before_filtering, excluding the 1st column (index, no use)
# static_house_updated=static_house_cleaned excluding some bldg_ids
# energy_bldg has the total energy usage for each bldg, derived from model_table
# energy_hour has the total and average energy usage for every hour, derived from model_table
# house_energy is the result of static_house_updated left join energy_bldg
```

## Read in data

I used this table excluding some bldg_ids but not excluding columns with low variances. I renamed it model_table for later usage

```{r Hang}
library(tidyverse)
library(lubridate)
# Setting working directory, mute it and set it to your working file
  # Shahukaru:
  # setwd("/Users/r.shahukaru/Desktop/IST 687/Project")
  # Hang:
#   setwd("~/Documents/1 Learning in US/Semester files Fall 2023/IST 687 Intro to DS/IST687_Project")
# load('descriptive.RData')
```

## Plots and descriptive analysis

```{r Hang}
colnames(model_table)
# remove 1st column X which is just index
# 3:26 electricity usage
# 27:29 fuel
# 30:39 gas
# 40:44 propane
# 45 total energy usage
```

1. Let's see the total July energy usage for each building

```{r Hang}
model_table$electricity=apply(model_table[,3:26],1,sum)
model_table$fuel=apply(model_table[,27:29],1,sum)
model_table$gas=apply(model_table[,30:39],1,sum)
model_table$propane=apply(model_table[,40:44],1,sum)
energy_bldg=model_table%>%group_by(bldg_id)%>%
  summarise(sum_energy=sum(total),
            sum_electricity=sum(electricity),
            sum_fuel=sum(fuel),
            sum_gas=sum(gas),
            sum_propane=sum(propane))
# Obviously there are bldgs with negative energy usage
# Here's the list of those bldgs 
energy_bldg=energy_bldg[,-7]
list_neg_energy=energy_bldg%>%filter(sum_energy<=0)
# 16 building within 15 unique counties
nrow(unique(energy_bldg$county))
# the total is 46 counties
house_energy=left_join(static_house_updated,energy_bldg,by='bldg_id')

ggplot(energy_bldg)+geom_boxplot(aes(,sum_energy))+theme_bw()+
  ylab('Total July energy usage / House')+
  ggtitle('Total July Energy Usage Distribution')
```

Result: There are 16 buildings within 15 unique counties having negative total energy usage, I cannot understand their situation and suggest we exclude those bldgs. The list is shown in *list_neg_energy*

2. Now hourly distribution of energy usage

```{r Abhishek}
model_table$time=model_table$time%>%as_datetime(tz='EST')
model_table$hour=hour(model_table$time)
energy_hour=model_table%>%group_by(hour)%>%
  summarise(sum_energy=sum(total),
            sum_electricity=sum(electricity),
            sum_fuel=sum(fuel),
            sum_gas=sum(gas),
            sum_propane=sum(propane),
            avg_energy=mean(total),
            avg_electricity=mean(electricity),
            avg_fuel=mean(fuel),
            avg_gas=mean(gas),
            avg_propane=mean(propane))

```

Result: It semes there are no fuel and propane consumption for all the records in this model_table.


3. plotting distribution of hourly usage records
```{r Hang}
ggplot(model_table%>%filter(total>0))+geom_boxplot(aes(hour,total,group=hour))
```

Result: There are many records that could be considered as outliers for every hour. There are hourly total of negative values, too, as seen in last chunk of code.

4.Total Energy Consumption in July (Abhishek)
```{r Abhishek}
model_data <- load('descriptive.RData')
library(ggplot2)
library(lubridate)

model_data$time <- ymd_hms(model_data$time)

# Filter data for the month of July
july_data <- model_data %>% 
  filter(month(time) == 7)

# Line plot for total energy consumption over time in July
ggplot(july_data, aes(x = time, y = total)) +
  geom_line() +
  labs(title = "Total Energy Consumption in July", x = "Date", y = "Energy Consumption (kWh)") +
  theme_minimal()
```

4.Comparision of energy sources
```{r Abhishek}
library(ggplot2)
library(reshape2)
long_data <- melt(model_table, measure.vars = c("electricity", "fuel", "gas", "propane"))

ggplot(long_data, aes(x = variable, y = value, fill = variable)) +
  geom_bar(stat = "identity", position = position_dodge()) +
  labs(x = "Category", y = "Value", title = "Comparison of Energy Sources") +
  theme_minimal()
```
5.Energy Consumption by County
```{r Abhishek}
# Load necessary libraries
library(dplyr)
library(ggplot2)
library(reshape2)

summarized_data <- model_table %>%
  group_by(county) %>%
  summarise(electricity = sum(electricity, na.rm = TRUE),
            fuel = sum(fuel, na.rm = TRUE),
            gas = sum(gas, na.rm = TRUE),
            propane = sum(propane, na.rm = TRUE))

long_data <- melt(summarized_data, id.vars = "county")

ggplot(long_data, aes(x = county, y = value, fill = variable)) +
  geom_bar(stat = "identity", position = position_dodge()) +
  labs(x = "County", y = "Total Consumption", title = "Energy Consumption by County") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1, size = 8)) # Rotate and adjust the size
```
5.Energy consumption by appliances
```{r Abhishek}
# Load necessary libraries
library(dplyr)
library(ggplot2)
library(tidyr)

appliance_sums <- model_table %>%
  summarise(
    CeilingFan = sum(out.electricity.ceiling_fan.energy_consumption, na.rm = TRUE),
    ClothesDryer = sum(out.electricity.clothes_dryer.energy_consumption, na.rm = TRUE),
    ClothesWasher = sum(out.electricity.clothes_washer.energy_consumption, na.rm = TRUE),
    CoolingFansPumps = sum(out.electricity.cooling_fans_pumps.energy_consumption, na.rm = TRUE),
    Cooling = sum(out.electricity.cooling.energy_consumption, na.rm = TRUE),
    Dishwasher = sum(out.electricity.dishwasher.energy_consumption, na.rm = TRUE),
    Freezer = sum(out.electricity.freezer.energy_consumption, na.rm = TRUE),
    HeatingFansPumps = sum(out.electricity.heating_fans_pumps.energy_consumption, na.rm = TRUE),
    HeatingHpBkup = sum(out.electricity.heating_hp_bkup.energy_consumption, na.rm = TRUE),
    Heating = sum(out.electricity.heating.energy_consumption, na.rm = TRUE),
    HotTubHeater = sum(out.electricity.hot_tub_heater.energy_consumption, na.rm = TRUE),
    HotTubPump = sum(out.electricity.hot_tub_pump.energy_consumption, na.rm = TRUE),
    HotWater = sum(out.electricity.hot_water.energy_consumption, na.rm = TRUE),
    LightingExterior = sum(out.electricity.lighting_exterior.energy_consumption, na.rm = TRUE),
    LightingGarage = sum(out.electricity.lighting_garage.energy_consumption, na.rm = TRUE),
    LightingInterior = sum(out.electricity.lighting_interior.energy_consumption, na.rm = TRUE),
    MechVent = sum(out.electricity.mech_vent.energy_consumption, na.rm = TRUE),
    PlugLoads = sum(out.electricity.plug_loads.energy_consumption, na.rm = TRUE),
    PoolHeater = sum(out.electricity.pool_heater.energy_consumption, na.rm = TRUE),
    PoolPump = sum(out.electricity.pool_pump.energy_consumption, na.rm = TRUE),
    Pv = sum(out.electricity.pv.energy_consumption, na.rm = TRUE),
    RangeOven = sum(out.electricity.range_oven.energy_consumption, na.rm = TRUE),
    Refrigerator = sum(out.electricity.refrigerator.energy_consumption, na.rm = TRUE),
    WellPump = sum(out.electricity.well_pump.energy_consumption, na.rm = TRUE)
  ) %>% 
  pivot_longer(cols = everything(), names_to = "Appliance", values_to = "Consumption")

# Plot the total electricity consumption for each appliance in descending order
ggplot(appliance_sums, aes(x = reorder(Appliance, -Consumption), y = Consumption)) +
  geom_bar(stat = "identity", fill = "blue") +
  labs(x = "Appliance", y = "Total Electricity Consumption (kWh)", title = "Total Electricity Consumption by Appliance") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) +
  guides(fill=FALSE)
```
